Add composer file for wordpress-theme install
Let's support composer installers for multi-framework support in wordpress.
By doing so, we have the ability to define a custom themes folder and bring this
as a definite dependency to build upon. This also makes it easier on
build/deploy processes where we don't have to worry about controlling a parent
theme folder that is being maintained by someone else.

For more info: https://github.com/composer/installers