Fix out-of-the-box behavior of mobile layout
When we just install the theme and do not change the mobile
layout and menu position in the customizer, these theme mods
return false.

Added handling for this so that when the mobile menu layout is
false, the off canvas will be shown; when the mobile menu position is
false, the left position will be used.