Added Size Class Back Into Image Tag Within Content
I agree with most of what this function is doing. However I needed to
apply styles to a particular image size within the body of a post and
realized that this function completely stripped out any unique
identifier that would have allowed me to apply styles to it. So I went
ahead and altered the function to include any classes prefixed by
“size”. This should allow developers to single out specific sizes
within posts and alter the accordingly.