Remove wp versions filter to support cdn caching
CDNs use version numbers to deliver new assets. When version number does not
exist, the asset is treated as the same and therefore must be flushed. Flushing
can be an issue because it creates greater responsibility in having to manage
clearing assets. It would be far easier for us to use verion number as default
to offer support to CDNs

A starting theme should not make the assumption that versioning of assets is not
needed.