fix sidebar search to use input-group
The sidebar should be using the correct foundation element "input-group"
instead of columns to display the search form. This fixes a disparity
betwen the text box and button heights, and makes the field take up the
full sidebar width. This also fixes an extra 15px padding on the sidebar
that shouldn't be there.