Changed search template to match default page mark up Removed hard coded css classes 'small-12 large-8 columns' which are a legacy feature from older versions.
Using the 'main-content' css class the page inherits the same behaviour as the default page.

Changed page header to h1 and placed inside a header tag similar to 'page.php'.

The purpose of these changes was to have the search page follow the same behaviour as the other pages in the template.